feat: implement user management endpoints for creation, retrieval, and listing
