+++ app/routes/users/__init__.py
@@ -2,4 +2,4 @@ from flask import Blueprint
 
 users = Blueprint('users', __name__)
 
-from . import routes
+from . import create_user, list_users, get_user
