+++ app/routes/users/create_user.py
@@ -3,32 +3,6 @@ from werkzeug.security import generate_password_hash
 from . import users
 from app.utils.db_connection import MySQLConnection
 
-@users.route('/')
-def list_users():
-    db = MySQLConnection()
-    try:
-        query = "SELECT id, username, email, role_id, created_at FROM users"
-        users = db.execute_query(query)
-        return jsonify(users), 200
-    except Exception as e:
-        return jsonify({'error': str(e)}), 500
-    finally:
-        db.close()
-
-@users.route('/<int:user_id>')
-def get_user(user_id):
-    db = MySQLConnection()
-    try:
-        query = "SELECT id, username, email, role_id, created_at FROM users WHERE id = %s"
-        user = db.execute_query(query, (user_id,))
-        if not user:
-            return jsonify({'error': 'User not found.'}), 404
-        return jsonify(user[0]), 200
-    except Exception as e:
-        return jsonify({'error': str(e)}), 500
-    finally:
-        db.close()
-
 @users.route('/', methods=['POST'])
 def create_user():
     db = MySQLConnection()
